updated code 11-3

+++ frontend/src/pages/Cor.jsx
@@ -19,18 +19,25 @@ const Cor = () => {
     // Handle logout
     const handleLogOut = async () => {
         try {
-            await axios.delete("http://localhost:3001/apiu/logout");
+          const userConfirmed = window.confirm("Do you want to proceed?");
+        
+          if (userConfirmed) {
+            axios.delete("http://localhost:3001/apiu/logout");
+            sessionStorage.clear();
             alert("Logged Out");
-            navigate('/');
+            navigate('/', { replace: true });
+          } else {
+            navigate('/cor') 
+          } 
         } catch (error) {
-            console.error("Logout error:", error);
-            alert("Failed to log out. Please try again.");
+          console.error("Error during logout:", error);
+          alert("Failed to log out. Please try again.");
         }
-    };
+      };
 
     // Navigate to home
     const goHome = () => {
-        navigate('/landing');
+        navigate('/landing', { replace: true });
     };
 
     // Fetch universities data
@@ -83,7 +90,7 @@ const Cor = () => {
                         </li>
                     </ul>
                 )}
-                <div className="div2">
+                <div className="div21">
                     <div className="grand-container">
                         <h1>VIEW COURSES</h1>
                         <table>
